[ADF-5580] Remove redundant setup and use ContentTestingModule in comment components tests
