+++ lib/content-services/src/lib/node-comments/node-comments.component.spec.ts
@@ -17,21 +17,10 @@
 
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { NodeCommentsComponent } from './node-comments.component';
-import { ADF_COMMENTS_SERVICE } from '../../../../core/src/lib/comments/interfaces/comments.token';
 import { CommentsComponent } from '../../../../core/src/lib/comments/comments.component';
 import { CommentModel } from '../../../../core/src/lib/models';
-import { NodeCommentsServiceMock } from './mocks/node-comments.service.mock';
 import { UnitTestingUtils } from '../../../../core/src/lib/testing/unit-testing-utils';
-import { provideHttpClient } from '@angular/common/http';
-import { AppConfigService } from '../../../../core/src/public-api';
-import { AppConfigServiceMock } from '../../../../core/src/lib/common/mock/app-config.service.mock';
-import { ContentService } from '../../../../content-services/src/lib/common/services/content.service';
-import { AuthenticationService } from '../../../../core/src/lib/auth/services/authentication.service';
-import { RedirectAuthService } from '../../../../core/src/lib/auth/oidc/redirect-auth.service';
-import { NoopTranslateModule } from '../../../../core/src/lib/testing/noop-translate.module';
-import { AlfrescoApiService } from '../services/alfresco-api.service';
-import { AlfrescoApiServiceMock } from '../mock/alfresco-api.service.mock';
-import { NoopAnimationsModule } from '@angular/platform-browser/animations';
+import { ContentTestingModule } from '../testing/content.testing.module';
 
 describe('NodeCommentsComponent', () => {
     let fixture: ComponentFixture<NodeCommentsComponent>;
@@ -40,34 +29,7 @@ describe('NodeCommentsComponent', () => {
 
     beforeEach(async () => {
         await TestBed.configureTestingModule({
-            imports: [NodeCommentsComponent, NodeCommentsComponent, NoopTranslateModule, NoopAnimationsModule],
-            providers: [
-                provideHttpClient(),
-                {
-                    provide: ADF_COMMENTS_SERVICE,
-                    useClass: NodeCommentsServiceMock
-                },
-                {
-                    provide: AppConfigService,
-                    useClass: AppConfigServiceMock
-                },
-                {
-                    provide: AlfrescoApiService,
-                    useClass: AlfrescoApiServiceMock
-                },
-                {
-                    provide: ContentService,
-                    useValue: {}
-                },
-                {
-                    provide: AuthenticationService,
-                    useValue: {}
-                },
-                {
-                    provide: RedirectAuthService,
-                    useValue: {}
-                }
-            ]
+            imports: [NodeCommentsComponent, ContentTestingModule]
         }).compileComponents();
 
         fixture = TestBed.createComponent(NodeCommentsComponent);
